common/requirements.txt
boto3
botocore
certifi
charset-normalizer
idna
jmespath
psycopg2-binary
python-dateutil
requests
s3transfer
six
urllib3
pandas
pydantic
fsspec
s3fs
